Backend/esp/requirements.txt
asgiref==3.3.1
certifi==2021.10.8
charset-normalizer==2.0.9
Django==3.1.5
django-environ==0.8.1
django-extensions==3.1.5
djangorestframework==3.12.2
idna==3.3
psycopg2==2.9.2
pytz==2020.5
requests==2.26.0
sqlparse==0.4.1
urllib3==1.26.7